requirements.txt
google-generativeai>=0.3.0
opencv-python>=4.5.0
pyaudio>=0.2.11
pillow>=8.0.0
mss>=6.1.0
PyQt5>=5.15.0
py2app>=0.28.0 
pynput>=1.7.0
python-dotenv>=0.19.0